+++ CHANGELOG.md
@@ -3,5 +3,5 @@ All notable changes to this project will be documented in this file. This change
 
 ## [0.1.0]
 ### Added
-- First version of **tasks** - milestones 1.0.0 and tasks 0.2.4
+- First version of **tasks** - milestones 1.0.1 and tasks 0.2.5